refactor: ShoppingUtils 헬퍼 클래스로 분리
