+++ app/src/main/java/woowacourse/shopping/view/payment/PaymentFragment.kt
@@ -12,8 +12,8 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.repository.remote.RemoteCouponRepositoryImpl
 import woowacourse.shopping.data.repository.remote.RemoteOrderRepositoryImpl
 import woowacourse.shopping.databinding.FragmentPaymentBinding
-import woowacourse.shopping.utils.ShoppingUtils.makeToast
 import woowacourse.shopping.utils.exception.ErrorEvent
+import woowacourse.shopping.utils.helper.ToastMessageHelper.makeToast
 import woowacourse.shopping.view.MainActivityListener
 import woowacourse.shopping.view.ViewModelFactory
 import woowacourse.shopping.view.cart.model.ShoppingCart
